API: Update lib.stride_tricks namespace
