+++ numpy/__init__.pyi
@@ -536,7 +536,7 @@ from numpy.lib.shape_base import (
     put_along_axis as put_along_axis,
 )
 
-from numpy.lib.stride_tricks import (
+from numpy.lib._stride_tricks_impl import (
     broadcast_to as broadcast_to,
     broadcast_arrays as broadcast_arrays,
     broadcast_shapes as broadcast_shapes,
