+++ numpy/core/_methods.py
@@ -82,7 +82,7 @@ def _count_reduce_items(arr, axis, keepdims=False, where=True):
         # axis and full sum is more excessive than needed.
 
         # guarded to protect circular imports
-        from numpy.lib.stride_tricks import broadcast_to
+        from numpy.lib._stride_tricks_impl import broadcast_to
         # count True values in (potentially broadcasted) boolean mask
         items = umr_sum(broadcast_to(where, arr.shape), axis, nt.intp, None,
                         keepdims)
